requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0

# Basic utilities
numpy>=1.24
tqdm>=4.66
pillow>=10.0
pandas>=2.0
PyYAML>=6.0
torch-geometric

# NLP / Text
sentence-transformers>=2.2.2
transformers>=4.36
ftfy>=6.1
regex>=2023.0.0
scikit-learn>=1.3

# Optional but useful
matplotlib>=3.7
